Refactor components
Create Form component, and refactor General & Name components so that
the Form component passes props to them. This includes separating the
form stylings into its own 'Form.css' file, and creating a
FormFields.jsx file that contains the form field data.

This is being done so that the overall form state can be better
controlled when changes are made to its smaller elements, i.e. inputs.
This will allow for proper page updates when the submit button is
clicked.

Right now this only updates the value of the 'name' input field upon
user changes. Next step is to add the checkValidity function to the
Form component, finish the handleChange logic, and add handleBlur.