Refactor / change submitted form logic
Remove all previously implemented logic to update the display when
there is a valid form submission. New display updates will be rebuilt
from scratch and put in an entirely new / separate component.

This includes moving much of the form state / function info from
Form.jsx to App.jsx. With the new logic, when the form status is
"submitted," it will not display the form at all, and rather present
the form data as other HTML elements. This is cleaner and more
semantically sound