Hide 'field of study' form field for diploma
Hide the Field of Study form field for the diploma / GED option, and
show it for every other degree option.

This includes adding an 'isPresent' property to the field of study
property in FormFields.jsx, as well as a fairly significant refactor of
App.jsx, which includes removing checkValidity as a separate function
and adding the logic in the handleChange and handleBlur functions. The
latter is done to properly control the form field data when the field
of study field is present/not present, and it has an added cleanliness
bonus, as there will no longer be any occurrence where setFormFields is
called multiple times within one state change.

This likely breaks proper form submission / display, so updating that
logic will be next to do.